work on todo-app input form

+++ to-do/todo-client.js
@@ -1,13 +1,14 @@
 function deleteItem () {
 
-  $.ajax({
-      type: "DELETE",
-      url: "localhost:3000",
-      data: "name=someValue",
-      success: function(msg){
-          alert("Data Deleted: " + msg);
-      }
-  });
-
+  var item = $("#inputText").val();
+  if (item !='')
+      $.ajax({
+          type: "DELETE",
+          url: "localhost:3000",
+          data: item,
+          success: function(msg){
+              alert("Data Deleted: " + msg);
+          }
+      });
 
 }
